refactor: class LottoDao
Dao에 들어있던 Connection 관련 메서드를 DataConnection 클래스로 분리
