+++ README.md
@@ -137,7 +137,7 @@
 2. post: `/lotto`
 
 index.html에서 입력받은 값으로 lotto를 진행
-한번에 모든 입력값을 `/lotto`로 보내 최종 결과값을 ajax로 얻어와 보여주는 구조
+한번에 모든 입력값을 `/lotto`로 보내 최종 결과값 `/result`로 보여주는 구조
 
 ## 로또 5단계 - DB 적용
