+++ src/test/java/lotto/dao/LottoDaoTest.java
@@ -5,7 +5,6 @@ import lotto.dto.LottoDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -15,19 +14,12 @@ import static lotto.domain.generator.LottoNumbersGenerator.generateLottoNumbers;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class LottoDaoTest {
-
     private static final int FIRST_ROUND = 1;
     LottoDao lottoDao;
 
     @BeforeEach
     void setUp() {
         lottoDao = new LottoDao();
-    }
-
-    @Test
-    void 연결_테스트() {
-        Connection con = lottoDao.getConnection();
-        assertThat(con).isNotNull();
     }
 
     @Test
